requirements.txt
PyMySQL==1.0.2
SQLAlchemy==1.3.23
fastapi==0.63.0
itsdangerous==1.1.0
mysql-connector==2.2.9
passlib==1.7.4
pydantic==1.7.3
python-multipart==0.0.5
starlette==0.13.6
uvicorn==0.13.3
python-jose==3.2.0
bcrypt==3.2.0
captcha==0.3